Building/Installation of pycompwa via pip
pycompwa can now easily be installed via pip. pycompwa are only python
modules and the python interface to ComPWA. ComPWA itself is introduced
via a git submodule. This submodule is build on the fly, if no existing
ComPWA installation can be found.
Also an appropriate travis CI was set up, which tests the installation
process with ComPWA and runs all tests of pycompwa.

* Added ComPWA as submodule
* Added pybind11 as submodule
* Use scikit-build to build ComPWA using cmake
* Updated __init__.py to load interface from other directory
* Added .travis.yml and license
* Moved tests folder to top level
* Added Pipfile and Pipfile.lock
* Added dalitzfit example