Fixed travis CI python build and test envs
The current pipenv setup broke on travis CI. Since pipenv is extremely
slow, I switched back to the standard python venv. However here the
dependency management for the test environment is not working correctly.
It was manually fixed and should be solved correctly with issue #37.
Consequently the Pipfiles were also removed.